(1) Renamed results class. (2) Moved JSON translation from service to gateway.
