+++ src/main/java/com/eames/masterkey/aws/gateway/http/BittingListHTTPGateway.java
@@ -2,11 +2,12 @@ package com.eames.masterkey.aws.gateway.http;
 
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestStreamHandler;
-import com.eames.masterkey.service.progression.ProgressionServiceException;
-import com.eames.masterkey.service.progression.ProgressionServiceProviderException;
-import com.eames.masterkey.service.progression.ProgressionService;
-import com.eames.masterkey.service.progression.ProgressionServiceProvider;
+import com.eames.masterkey.service.progression.*;
 import com.eames.masterkey.service.progression.services.totalposition.RandomGenericTotalPositionProgressionService;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.JsonPrimitive;
+import com.google.gson.JsonSerializer;
 import org.apache.commons.io.IOUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -77,12 +78,28 @@ public class BittingListHTTPGateway
 
             // Generate the bitting list.
             // Throws: ProgressionServiceException
-            String responseStr = service.generateBittingList(jsonRequestStr);
-            logger.debug("Bitting List: {}", responseStr);
+            ProgressionServiceResults results = service.generateBittingList(jsonRequestStr);
+
+            /*
+             * Construct a gson instance using the gson builder.
+             * Specify that int arrays should be serialized as strings.
+             */
+            Gson gson = new GsonBuilder()
+                    .registerTypeAdapter(int[].class, (JsonSerializer<int[]>) (src, type, jsonSerializationContext) -> {
+                        StringBuilder sb = new StringBuilder();
+                        for (int v : src)
+                            sb.append(v);
+                        return new JsonPrimitive(sb.toString());
+                    })
+                    .create();
+
+            // Serialize the results to JSON.
+            String resultsStr = gson.toJson(results);
+            logger.debug("Results: {}", resultsStr);
 
             // Write the response to the output stream.
             OutputStreamWriter writer = new OutputStreamWriter(outputStream, "UTF-8");
-            writer.write(responseStr);
+            writer.write(resultsStr);
             writer.close();
 
         } catch (ProgressionServiceProviderException ex) {
