+++ src/main/java/com/eames/masterkey/service/progression/ProgressionCriteria.java
@@ -1,0 +1,7 @@
+package com.eames.masterkey.service.progression;
+
+/**
+ * This interface is a marker interface for all progression service criteria.
+ */
+public interface ProgressionCriteria {
+}
